extract all remaining reductions
